\name{eatlessmeat-package}
\alias{eatlessmeat-package}
\alias{eatlessmeat}
\docType{package}
\title{ 
Eat less meat, save water?
}
\description{
Water footprint analysis of effect of eating less meat. Uses national diet information to identify 
the minimal changes to a diet to meet nutritional requirements while reducing meat consumption. Uses
national water footprint data to quantify the resulting reduction in water footprint.
}
\author{
Joseph Guillaume, based on work by Mika Jalava et al.
Maintainer: Joseph Guillaume <josephguillaume@gmail.com>
}
\references{
Jalava, M, M Kummu, M Porkka, S Siebert, and O Varis. 2014. “Diet Change—a Solution to Reduce Water Use?” Environmental Research Letters 9 (7). IOP Publishing: 074016. doi:10.1088/1748-9326/9/7/074016.
Original Matlab implementation available at: http://iopscience.iop.org/1748-9326/9/7/074016/media
}
\keyword{ package }